kitti is training

+++ core/corr.py
@@ -37,12 +37,11 @@ class CorrBlock:
             dx = torch.linspace(-r, r, 2*r+1)
             dy = torch.linspace(-r, r, 2*r+1)
             delta = torch.stack(torch.meshgrid(dy, dx), axis=-1).to(coords.device)
-
             centroid_lvl = coords.reshape(batch*h1*w1, 1, 1, 2) / 2**i
             delta_lvl = delta.view(1, 2*r+1, 2*r+1, 2)
             coords_lvl = centroid_lvl + delta_lvl
 
-            corr = bilinear_sampler(corr, coords_lvl)
+            corr = bilinear_sampler(corr, coords_lvl.type( corr.dtype) )
             corr = corr.view(batch, h1, w1, -1)
             out_pyramid.append(corr)
